Add ability to specify the path to a local configuration file to lsc
Example:
lsc --config=~/path/to/local-config.json

The values from the config file are loaded first. Configuration values
in the current directory's config.json are loaded second. This change
allows config files to be stored in a central location instead of forcing
them to be stored in the root directory of LabShare packages.

 - Update unit tests
 - Minor refactoring